refactor #127 : 컴포넌트명 구체화
