+++ src/common/components/layout/header-buttons/total-evaluation/total-evaluation-sidebar-buttons.tsx
@@ -4,9 +4,9 @@ import Icon from '@/common/components/icon/icon';
 import { SidebarContext } from '@/features/total-evaluation/components/context/sidebar/sidebar-context';
 import { SidebarOpenButton } from '@/features/total-evaluation/components/custom-buttons/sidebar-open-button';
 
-import * as styles from './total-evaluation-left-buttons.styles';
+import * as styles from './total-evaluation-sidebar-buttons.styles';
 
-function TotalEvalutationLeftButtons() {
+function TotalEvalutationSidebarButtons() {
   const { setIsSidebarOpen } = useContext(SidebarContext);
   return (
     <div css={styles.container}>
@@ -16,4 +16,4 @@ function TotalEvalutationLeftButtons() {
   );
 }
 
-export default TotalEvalutationLeftButtons;
+export default TotalEvalutationSidebarButtons;
